Add path[][] texpath(string s) routine to convert string into the paths that TeX would fill.

+++ picture.cc
@@ -330,7 +330,8 @@ bool picture::postprocess(const string& prename, const string& outname,
 	
       if(running) {
 	if(Viewer == "gv") kill(pid,SIGHUP); // Tell gv to reread file.
-	else if(pdfformat) { // Kill pdfviewer so that file can be redrawn.
+	else if(pdfformat && !b.empty) {
+	  // Kill pdfviewer so that file can be redrawn.
 	  kill(pid,SIGINT);
 	  while(waitpid(pid, &status, 0) != pid);
 	  running=false;
